impl: extract write-form component with custom callback & handle submit api response
