+++ apps/web/app/layout.tsx
@@ -3,6 +3,7 @@ import { ColorSchemeScript, MantineProvider, createTheme } from "@mantine/core";
 import { Notifications } from "@mantine/notifications";
 import "@mantine/core/styles.css";
 import "@mantine/notifications/styles.css";
+import "@mantine/tiptap/styles.css";
 import type { Metadata } from "next";
 import "./globals.css";
 
@@ -30,7 +31,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }):
 			</head>
 			<body>
 				<MantineProvider theme={theme} defaultColorScheme="dark">
-					<Notifications />
+					<Notifications position="bottom-center" />
 					<ApplicationShell>{children}</ApplicationShell>
 				</MantineProvider>
 			</body>
