+++ apps/web/types/errors.ts
@@ -3,5 +3,6 @@ export interface ErrorResponse {
 	status: number;
 }
 
+export const ERR_MISSING_TOKEN = "missing or malformed JWT";
 export const ERR_EXPIRED_TOKEN = "token has invalid claims: token is expired";
 export const ERR_REVIEW_EXIST = "You have already written a review for this course";
